excelent

+++ src/main/kotlin/ch/mofobo/impacct/dtos/PieChartData.kt
@@ -1,3 +1,3 @@
 package ch.mofobo.impacct.dtos
 
-data class PieChartData(val label: String, val value: Int)
+data class PieChartData(val label: String, val value: Float)
